+++ tlm_adjoint/__init__.py
@@ -21,7 +21,6 @@
 from .caches import *                # noqa: F401
 from .eigendecomposition import *    # noqa: F401
 from .equations import *             # noqa: F401
-from .float import *                 # noqa: F401
 from .functional import *            # noqa: F401
 from .hessian import *               # noqa: F401
 from .hessian import GeneralHessian as Hessian  # noqa: F401
@@ -29,5 +28,6 @@ from .hessian_optimization import *  # noqa: F401
 from .interface import *             # noqa: F401
 from .manager import *               # noqa: F401
 from .optimization import *          # noqa: F401
+from .overloaded_float import *      # noqa: F401
 from .tlm_adjoint import *           # noqa: F401
 from .verification import *          # noqa: F401
